Completed Final Project - Part 2
Rewrote/reformatted some wording in the manuscript. Cleaned and explored the data, creating plots of distributions and correlations, and then added them to the manuscript. Removed files that did not go with my project/from the template. Made sure it was reproducible and was able to knit into an html.